requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
comm==0.2.3
configobj==5.0.9
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.63.0
dvc-data==3.16.12
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
entrypoints==0.4
executing==2.2.1
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
funcy==2.0
gitdb==4.0.12
GitPython==3.1.45
grandalf==0.8
gto==1.8.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
iterative-telemetry==0.0.10
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
kombu==5.5.4
lazy_loader==0.4
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
orjson==3.11.3
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.5
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scipy==1.16.1
scmrepo==3.5.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sqltrie==0.11.2
stack-data==0.6.3
sympy==1.14.0
tabulate==0.9.0
tifffile==2025.8.28
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
yarl==1.20.1
zc.lockfile==3.0.post1
